tests/test_fetch_emails.py
import unittest
from unittest.mock import MagicMock, patch
from email_manager.fetch_emails import EmailFetcher

class TestEmailFetcher(unittest.TestCase):
    @patch('email_manager.fetch_emails.GmailAuthenticator')
    @patch('email_manager.fetch_emails.get_session')
    def test_fetch_emails_success(self, mock_get_session, mock_GmailAuthenticator):
        mock_service = MagicMock()
        mock_service.users().messages().list().execute.return_value = {'messages': [{'id': '1'}]}
        mock_service.users().messages().get().execute.return_value = {
            'payload': {'headers': [{'name': 'From', 'value': 'test@example.com'},
                                    {'name': 'Subject', 'value': 'Test Subject'}]},
            'internalDate': '1622015746000'
        }
        mock_GmailAuthenticator().authenticate.return_value = mock_service
        email_fetcher = EmailFetcher()
        email_fetcher.fetch_emails()
        # Add assertions here to validate the fetched emails and their insertion into the database

    @patch('email_manager.fetch_emails.GmailAuthenticator')
    @patch('email_manager.fetch_emails.get_session')
    def test_fetch_emails_exception_handling(self, mock_get_session, mock_GmailAuthenticator):
        mock_service = MagicMock()
        mock_service.users().messages().list().execute.side_effect = Exception('API Error')
        mock_GmailAuthenticator().authenticate.return_value = mock_service
        email_fetcher = EmailFetcher()
        with self.assertRaises(Exception):
            email_fetcher.fetch_emails()
        # Add assertions here to validate the error handling/logging

if __name__ == '__main__':
    unittest.main()
